move tests in a separate sub-module in order to avoid the global() related hacks performed in install_tests_in_module_dict()
